+++ Magic/CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Add /magic help command, simple command that relies on /help for more info
  - Add heroes_skill_prefix config property, to prefix skill names from the Magic Heroes skill pack
+ - Add /mitem damage command
 
 ## 6.7.1